backend/requirements.txt
# FastAPI & server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Face recognition / deep learning
facenet-pytorch==2.6.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.10.0.82

# Encryption
pycryptodome==3.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.7.4
pydantic-settings==2.3.3
scipy==1.13.1
